Rename sky proxy to sky forwarding
