+++ cmd/skywire-cli/commands/root.go
@@ -19,7 +19,7 @@ import (
 	climdisc "github.com/skycoin/skywire/cmd/skywire-cli/commands/mdisc"
 	clireward "github.com/skycoin/skywire/cmd/skywire-cli/commands/reward"
 	clirtfind "github.com/skycoin/skywire/cmd/skywire-cli/commands/rtfind"
-	cliskyproxy "github.com/skycoin/skywire/cmd/skywire-cli/commands/skyproxy"
+	cliskyforwarding "github.com/skycoin/skywire/cmd/skywire-cli/commands/skyforwarding"
 	clisurvey "github.com/skycoin/skywire/cmd/skywire-cli/commands/survey"
 	clivisor "github.com/skycoin/skywire/cmd/skywire-cli/commands/visor"
 	clivpn "github.com/skycoin/skywire/cmd/skywire-cli/commands/vpn"
@@ -175,7 +175,7 @@ func init() {
 		clirtfind.RootCmd,
 		climdisc.RootCmd,
 		clicompletion.RootCmd,
-		cliskyproxy.RootCmd,
+		cliskyforwarding.RootCmd,
 		treeCmd,
 		docCmd,
 	)
